DAO usuarios

+++ GestorDeTareas/src/main/java/org/example/modelo/Categoria.java
@@ -4,19 +4,19 @@ import java.util.List;
 import src.main.java.org.example.modelo.Tareas;
 
 public class Categoria {
-    private int id;
+    private Long id;
     private String nombre;
     private String descripcion;
     private List<Tareas> tareas;
-    public Categoria(int id, String nombre, String descripcion) {
+    public Categoria(Long id, String nombre, String descripcion) {
         this.id = id;
         this.nombre = nombre;
         this.descripcion = descripcion;
     }
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getNombre() {
